Fix libmesh domain partitioning
This was a slightly confusing one, it seemed that when doing the OpenFOAM
to MOOSE transfer, we could set values on the MOOSE mesh fine using some
element number and DoF. However, using that same element number and DoF
to get a value would cause an access violation when running the app in
parallel.

It appears it came down to the libmesh mesh not having the same domain
decomposition as the OpenFOAM mesh, as we weren't initialising it
properly. It still confuses me a bit that an MPI rank can set values on
the mesh outside of its domain, I would have expected a similar access
violation error.