Add input file for hippo part of heated plate test
This models the solid heated plate using the heat conduction kernel in
MOOSE. The idea is that this input file will be run as part of a
hippo multiapp, where an OpenFOAM fluid will run over the top of the
heated plate. The plate is heated from the bottom and will begin
transferring heat to the fluid.