Implement 'warm_plate' multiapp test
This is a simplified version of the 'heated plate' test. It made it
clear that we needed to reverse the direction (negate) of the heat flux
when setting it as a boundary condition in MOOSE.

The new test has no fluid flow and both the fluid and solid domains have
no boundary conditions (other than at their interface), and have a
uniform temperature.

We need to do more testing of this example to check that we're
conserving all the quantities we need to (is it just energy we need to
consider?).